Repackaged Code to include Logic [command, parser] and model [modules and modulelist]
