+++ src/main/java/seedu/igraduate/logic/command/DeleteCommand.java
@@ -1,15 +1,15 @@
-package seedu.igraduate.command;
+package seedu.igraduate.logic.command;
 
-import seedu.igraduate.Storage;
-import seedu.igraduate.ModuleList;
-import seedu.igraduate.Ui;
+import seedu.igraduate.storage.Storage;
+import seedu.igraduate.model.ModuleList;
+import seedu.igraduate.ui.Ui;
 
 import seedu.igraduate.exception.ModuleNotFoundException;
 import seedu.igraduate.exception.PrerequisiteNotFoundException;
 import seedu.igraduate.exception.SaveModuleFailException;
 import seedu.igraduate.exception.UnableToDeletePrereqModuleException;
 
-import seedu.igraduate.module.Module;
+import seedu.igraduate.model.module.Module;
 
 import java.util.logging.Logger;
 import java.util.logging.Level;
