+++ src/test/java/seedu/igraduate/IGraduateTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.Test;
 
 import seedu.igraduate.exception.InvalidCommandException;
+import seedu.igraduate.logic.Parser;
 
 class IGraduateTest {
     @Test
