+++ src/test/java/seedu/igraduate/ModuleListTest.java
@@ -8,7 +8,8 @@ import seedu.igraduate.exception.ExistingModuleException;
 import seedu.igraduate.exception.ModuleNotFoundException;
 import seedu.igraduate.exception.PrerequisiteNotFoundException;
 import seedu.igraduate.exception.UnableToDeletePrereqModuleException;
-import seedu.igraduate.module.GeModule;
+import seedu.igraduate.model.ModuleList;
+import seedu.igraduate.model.module.GeModule;
 
 import java.util.ArrayList;
