+++ src/test/java/seedu/igraduate/logic/command/CapCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.igraduate.command;
+package seedu.igraduate.logic.command;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -7,10 +7,10 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.igraduate.ModuleList;
-import seedu.igraduate.Parser;
-import seedu.igraduate.Storage;
-import seedu.igraduate.Ui;
+import seedu.igraduate.model.ModuleList;
+import seedu.igraduate.logic.Parser;
+import seedu.igraduate.storage.Storage;
+import seedu.igraduate.ui.Ui;
 
 import seedu.igraduate.exception.InvalidModuleGradeException;
 import seedu.igraduate.exception.UnableToDeletePrereqModuleException;
